requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
google-cloud-aiplatform==1.38.1
google-cloud-firestore==2.14.0
google-auth==2.27.0
google-api-python-client==2.116.0
PyMuPDF==1.23.21
httpx==0.26.0
python-multipart==0.0.6
